+++ stories/primer/auto_complete_component_stories.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+class Primer::AutoCompleteComponentStories < ViewComponent::Storybook::Stories
+  layout "storybook_preview"
+
+  story(:auto_complete) do
+    controls do
+      text(:src, "/")
+      text(:id, "id")
+    end
+
+    content do |c|
+      c.input(name: "username")
+      c.icon(icon: :search)
+    end
+  end
+end