+++ test/components/auto_complete_item_component_test.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class PrimerAutoCompleteItemComponentTest < Minitest::Test
+  include Primer::ComponentTestHelpers
+
+  def test_renders_item_with_correct_attributes
+    render_inline(Primer::AutoCompleteItemComponent.new(value: "foo")) { "Item" }
+
+    assert_selector("li[role=\"option\"][data-autocomplete-value=\"foo\"].autocomplete-item", text: "Item")
+  end
+
+  def test_renders_selected_item
+    render_inline(Primer::AutoCompleteItemComponent.new(value: "foo", selected: true)) { "Item" }
+
+    assert_selector("li[role=\"option\"][data-autocomplete-value=\"foo\"][aria-selected=\"true\"].autocomplete-item", text: "Item")
+  end
+
+  def test_renders_disabled_item
+    render_inline(Primer::AutoCompleteItemComponent.new(value: "foo", disabled: true)) { "Item" }
+
+    assert_selector("li[role=\"option\"][data-autocomplete-value=\"foo\"][aria-disabled=\"true\"].autocomplete-item.disabled", text: "Item")
+  end
+end